change in upload

+++ src/utils/cloudinary.js
@@ -20,9 +20,18 @@ const uploadOnCloudinary = async (localFilePath) => {
         // file has been uploaded successfull
         console.log("file is uploaded on cloudinary ", response.url);
 
-        fs.unlinkSync(localFilePath);
+        // remove the locally saved temporary file as the upload operation got successfull
+
         
+    // Remove local file after successful upload
+       if (fs.existsSync(localFilePath)) {
+           fs.unlinkSync(localFilePath);
+        }
+
+
         return response;
+
+         
 
     } catch (error) {
         fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
